microblog/requirements.txt
alembic==1.15.2
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
elastic-transport==8.17.1
elasticsearch==8.17.2
email_validator==2.2.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psycopg==3.2.3
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rq==2.3.2
ruff==0.11.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
